refactor helper functions to make sonarcloud pass
